refactor(ui): separate session capture from restore
Move snapshot serialization, preservation merging, debounced persistence, and native flush hooks into useAppSessionCapture. Keep useAppSessionRestore focused on reconstructing workspace and SideCar tabs while retaining the same public hook in App.

Preserve the original startup ordering by preparing capture, releasing the restore gate, and only then activating reactive persistence. This keeps partial-restore state and shutdown flush behavior unchanged.

Validated with the UI typecheck, 58 focused restoration and client-state tests, the production UI build, whitespace checks, and a targeted lifecycle review.
